PPSS/ppss-2023-g10-asensi-roch-master/P05-Dependencias2/P05-mocks/reservaMocks/src/main/java/ppss/Reserva.java
package ppss;
import ppss.excepciones.*;

import java.util.ArrayList;
public class Reserva {

    private FactoriaBOs factoria = new FactoriaBOs();

    public void setFactoria(FactoriaBOs fac){
        factoria = fac;
    }
    public boolean compruebaPermisos(String login, String password, Usuario tipoUsu) {
        throw new UnsupportedOperationException("Not yet implemented");
    }

    public void realizaReserva(String login, String password,
                               String socio, String [] isbns) throws ReservaException {
        ArrayList<String> errores = new ArrayList<String>();

        //Dependencia PartialMock
        if(!compruebaPermisos(login, password, Usuario.BIBLIOTECARIO)) {
            errores.add("ERROR de permisos");
        } else {
            //Dependencia
            FactoriaBOs fd = factoria;
            IOperacionBO io = fd.getOperacionBO();
            try {
                for(String isbn: isbns) {
                    try {
                        //Dependencia
                        io.operacionReserva(socio, isbn);
                    } catch (IsbnInvalidoException iie) {
                        errores.add("ISBN invalido" + ":" + isbn);
                    }
                }
            } catch (SocioInvalidoException sie) {
                errores.add("SOCIO invalido");
            } catch (JDBCException je) {
                errores.add("CONEXION invalida");
            }
        }
        if (errores.size() > 0) {
            String mensajeError = "";
            for(String error: errores) {
                mensajeError += error + "; ";
            }
            throw new ReservaException(mensajeError);
        }
    }
}
